refactor: update package structure to hwr.oop.projects.peakpoker
