--- src/main/kotlin/hwr/oop/core/Card.kt
@@ -1,3 +1,0 @@
-package hwr.oop.core
-
-data class Card(val suit: String, val rank: String)
