--- src/main/kotlin/hwr/oop/core/Player.kt
@@ -1,7 +1,0 @@
-package hwr.oop.core
-
-class Player(
-    val name: String,
-){
-// TODO: Player needs to be implemented
-}
