+++ src/main/kotlin/hwr/oop/projects/peakpoker/Main.kt
@@ -1,9 +1,9 @@
-package hwr.oop
+package hwr.oop.projects.peakpoker
 
 import com.github.ajalt.clikt.core.main
 import com.github.ajalt.clikt.core.subcommands
-import hwr.oop.commands.InitCommand
-import hwr.oop.commands.PokerCommand
+import hwr.oop.projects.peakpoker.commands.InitCommand
+import hwr.oop.projects.peakpoker.commands.PokerCommand
 
 fun main(args: Array<String>) = PokerCommand()
     .subcommands(
