+++ src/main/kotlin/hwr/oop/projects/peakpoker/core/Card.kt
@@ -1,0 +1,3 @@
+package hwr.oop.projects.peakpoker.core
+
+data class Card(val suit: String, val rank: String)
