+++ src/main/kotlin/hwr/oop/projects/peakpoker/core/Player.kt
@@ -1,0 +1,7 @@
+package hwr.oop.projects.peakpoker.core
+
+class Player(
+    val name: String,
+){
+// TODO: Player needs to be implemented
+}
