+++ src/test/kotlin/hwr/oop/projects/peakpoker/CommandsTest.kt
@@ -1,7 +1,7 @@
-package hwr.oop
+package hwr.oop.projects.peakpoker
 
 import com.github.ajalt.clikt.testing.test
-import hwr.oop.commands.InitCommand
+import hwr.oop.projects.peakpoker.commands.InitCommand
 import io.kotest.core.spec.style.AnnotationSpec
 import org.assertj.core.api.Assertions.assertThat
